gateway-service/requirements.txt
fastapi
uvicorn[standard]
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp-proto-http
python-multipart